+++ ui/app/src/components/AwaitingApprovalDashletGrid/AwaitingApprovalDashletGridUI.tsx
@@ -116,7 +116,7 @@ export default function AwaitingApprovalDashletGridUI(props: AwaitingApprovalDas
                 </GlobalAppGridCell>
               </GlobalAppGridRow>
               <GlobalAppGridRow className="hoverDisabled">
-                <GlobalAppGridCell colSpan={7} className="padded0">
+                <GlobalAppGridCell colSpan={7} className="padded0 borderless">
                   <Collapse in={expandedLookup[dashboardItem.path]}>
                     <Table size="small" className={classes.tableRoot}>
                       <TableBody>
